+++ Tickers/Tickers/Scenes/Sounds/SoundModel.swift
@@ -1,0 +1,13 @@
+//
+//  SoundModel.swift
+//  Tickers
+//
+//  Created by Oliver Santos on 30/06/23.
+//
+
+struct Sound: Hashable {
+    let id: Int
+    let title: String
+    var isPlay: Bool = false
+    var isDownload: Bool
+}
