+++ Tickers/Tickers/Scenes/Timer/PomodoroView.swift
@@ -9,10 +9,10 @@ import SwiftUI
 
 struct PomodoroView: View {
     @State var sounds: [Sound] = [
-        Sound(title: "Chuva", isPlay: false, isDownload: false),
-        Sound(title: "Tempestade", isPlay: true, isDownload: true),
-        Sound(title: "Água Corrente", isPlay: true, isDownload: true),
-        Sound(title: "Lo-fi", isPlay: true, isDownload: true)
+        Sound(id: 0, title: "Chuva", isDownload: false),
+        Sound(id: 1, title: "Tempestade", isDownload: true),
+        Sound(id: 2, title: "Água Corrente", isDownload: true),
+        Sound(id: 3, title: "Lo-fi", isDownload: true)
     ]
     
     @State private var isPresented = false
